+++ src/main/java/me/coonect/coonect/member/application/domain/service/NicknameValidationService.java
@@ -3,7 +3,7 @@ package me.coonect.coonect.member.application.domain.service;
 import lombok.RequiredArgsConstructor;
 import me.coonect.coonect.member.application.port.in.NicknameValidationQuery;
 import me.coonect.coonect.member.application.port.in.NicknameValidationUseCase;
-import me.coonect.coonect.member.application.port.out.MemberRepository;
+import me.coonect.coonect.member.application.port.out.persistence.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
