+++ src/test/java/me/coonect/coonect/member/adapter/in/web/MemberResponseTest.java
@@ -1,8 +1,9 @@
-package me.coonect.coonect.member.adapter.in;
+package me.coonect.coonect.member.adapter.in.web;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
+import me.coonect.coonect.member.adapter.in.web.MemberResponse;
 import me.coonect.coonect.member.application.domain.model.Member;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
